Refacto: Moved events testing into dedicated events crate.
- cleaned up test_
- created sdl_binding module in events crate.
- updated sdl version
- commented out old events crate (not deleted waiting for confirmation)
