+++ crates/engine_core/src/lib.rs
@@ -1,9 +1,9 @@
-pub use events::events::*;
-use log::info;
+// pub use events::sdl_binding::*;
+// use log::info;
 
-#[cfg(not(feature = "override"))]
-pub fn entry() -> () {
-    info!("Loop");
-    events::events::apply_all();
-    ()
-}
+// #[cfg(not(feature = "override"))]
+// pub fn entry() -> () {
+//     info!("Loop");
+//     events::events::apply_all();
+//     ()
+// }
