+++ crates/events/src/lib.rs
@@ -1,35 +1,38 @@
-use log::info;
-use std::sync::Mutex;
+pub mod sdl_binding;
 
-pub mod event;
-pub mod sdl;
 
-pub static ALL_EVENTS: Mutex<Vec<fn() -> i64>> = Mutex::new(Vec::new());
+// use log::info;
+// use std::sync::Mutex;
 
-#[cfg(not(feature = "override"))]
-pub mod events {
+// pub mod event;
+// pub mod sdl;
 
-    use super::*;
+// pub static ALL_EVENTS: Mutex<Vec<fn() -> i64>> = Mutex::new(Vec::new());
 
-    #[cfg(not(feature = "override"))]
-    pub fn add(event: fn() -> i64) -> () {
-        let mut events = ALL_EVENTS.lock().unwrap();
-        events.push(event);
-        let name = std::any::type_name_of_val(&event);
-        info!("{name} event added");
-        ()
-    }
+// #[cfg(not(feature = "override"))]
+// pub mod events {
 
-    #[cfg(not(feature = "override"))]
-    pub fn remove(index: usize) -> () {
-        let mut events = ALL_EVENTS.lock().unwrap();
-        events.remove(index);
-        ()
-    }
+//     use super::*;
 
-    #[cfg(not(feature = "override"))]
-    pub fn apply_all() -> Vec<i64> {
-        let events = ALL_EVENTS.lock().unwrap();
-        events.iter().map(|f| f()).collect()
-    }
-}
+//     #[cfg(not(feature = "override"))]
+//     pub fn add(event: fn() -> i64) -> () {
+//         let mut events = ALL_EVENTS.lock().unwrap();
+//         events.push(event);
+//         let name = std::any::type_name_of_val(&event);
+//         info!("{name} event added");
+//         ()
+//     }
+
+//     #[cfg(not(feature = "override"))]
+//     pub fn remove(index: usize) -> () {
+//         let mut events = ALL_EVENTS.lock().unwrap();
+//         events.remove(index);
+//         ()
+//     }
+
+//     #[cfg(not(feature = "override"))]
+//     pub fn apply_all() -> Vec<i64> {
+//         let events = ALL_EVENTS.lock().unwrap();
+//         events.iter().map(|f| f()).collect()
+//     }
+// }
