+++ crates/events/src/sdl_binding/mod.rs
@@ -1,0 +1,4 @@
+mod app_events;
+mod input_events;
+mod window_events;
+pub mod event_manager;
