+++ crates/utils/Cargo.toml
@@ -4,6 +4,5 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-sdl3 = "0.16.2"
-
 events = { path = "../events" }
+sdl3 = "=0.17.3"
