+++ test_/Cargo.toml
@@ -6,4 +6,4 @@ edition = "2024"
 [dependencies]
 bitflags = "2.10.0"
 num_enum = "0.7.5"
-sdl3 = { version = "0.17.0", features= ["build-from-source"]}
+sdl3 = "=0.17.3"
